Rename the docs ports module to IO and say checksum file, not sidecar
"Ports" collides with base/common/ports.ts and base/node/ports.ts, which are
about TCP ports, and the module does HTTP - so the name invites the wrong
reading. "Sidecar" now reads as the container pattern more than the companion
file it means here; "checksum file" says the same thing plainly and pairs with
the digest it holds.
